Fixing 'bug' in filter function
Filter fn required context to be computed because it looped over node
inputs. This means that in compileGraph() even though it calls
computeContext() - it doesn't work because it walks the graph from
outputs backwards, and there aren't inputs to walk.

The fn still won't work for inputs filtering, as mentioned in the
comment, until context is computed.

Maybe hard coding inputs onto nodes that are saved will solve this
later.